+++ cropclassification/helpers/config_helper.py
@@ -42,7 +42,7 @@ class SensorData:
         if bands is not None:
             self.bands = bands
         else:
-            self.bands = self.imageprofile.bands
+            self.bands = self.imageprofile.bands  # type: ignore[assignment]
 
 
 def read_config(config_paths: List[Path], default_basedir: Optional[Path] = None):
@@ -171,7 +171,7 @@ def parse_sensordata_to_use(input) -> Dict[str, SensorData]:
 
 
 def _get_image_profiles(image_profiles_path: Path) -> Dict[str, ImageProfile]:
-    # Cropclassification gives best results with time_dimension_reducer "mean" for both
+    # Cropclassification gives best results with time_reducer "mean" for both
     # sentinel 2 and sentinel 1 images.
     # Init
     if not image_profiles_path.exists():
@@ -199,6 +199,7 @@ def _get_image_profiles(image_profiles_path: Path) -> Dict[str, ImageProfile]:
             base_image_profile=profiles_config[profile].get("base_image_profile"),
             collection=profiles_config[profile].get("collection"),
             bands=profiles_config[profile].getlist("bands"),
+            time_reducer=profiles_config[profile].get("time_reducer"),
             max_cloud_cover=profiles_config[profile].getfloat("max_cloud_cover"),
             process_options=profiles_config[profile].getdict("process_options"),
             job_options=profiles_config[profile].getdict("job_options"),
@@ -213,7 +214,7 @@ def _get_image_profiles(image_profiles_path: Path) -> Dict[str, ImageProfile]:
 def _validate_image_profiles(profiles: Dict[str, ImageProfile]):
     # Check that all base_image_profile s are actually existing image profiles.
     for profile in profiles:
-        base_image_profile = profiles[profile].base_image_profile
+        base_image_profile = profiles[profile].base_imageprofile
         if base_image_profile is not None and base_image_profile not in profiles:
             raise ValueError(
                 f"{base_image_profile=} not found for profile {profiles[profile]}"