+++ cropclassification/preprocess/_timeseries_helper.py
@@ -447,9 +447,9 @@ def calculate_periodic_timeseries(
                         period_band_data_df[statistic_columns_dict["std"]], axis=1
                     )
                     # Number of Files used
-                    period_band_data_df[
-                        f"{column_basename}_used_files"
-                    ] = period_band_data_df[statistic_columns_dict["max"]].count(axis=1)
+                    period_band_data_df[f"{column_basename}_used_files"] = (
+                        period_band_data_df[statistic_columns_dict["max"]].count(axis=1)
+                    )
 
                     # Only keep the columns we want to keep
                     columns_to_keep = [
@@ -517,7 +517,7 @@ def get_fileinfo_timeseries(path: Path) -> dict:
         imageinfo_values = imageinfo_part.split("_")
         orbit = None
         image_profile = None
-        time_dimension_reducer = None
+        time_reducer = None
         if "-" in imageinfo_values[0]:
             # OpenEO mosaic filename format
             image_profile = imageinfo_values[0]
@@ -537,7 +537,7 @@ def get_fileinfo_timeseries(path: Path) -> dict:
 
             start_date = datetime.fromisoformat(imageinfo_values[1])
             end_date = datetime.fromisoformat(imageinfo_values[2])
-            time_dimension_reducer = imageinfo_values[4]
+            time_reducer = imageinfo_values[4]
             band = imageinfo_values[-1]  # =last value
 
         else:
@@ -597,8 +597,8 @@ def get_fileinfo_timeseries(path: Path) -> dict:
             "band": band,
             "orbit": orbit,  # ASC/DESC
         }
-        if time_dimension_reducer is not None:
-            image_metadata["time_dimension_reducer"] = time_dimension_reducer
+        if time_reducer is not None:
+            image_metadata["time_reducer"] = time_reducer
         if image_profile is not None:
             image_metadata["image_profile"] = image_profile